+++ src/main/java/soma/edupi/gateway/EdupiGatewayApplication.java
@@ -1,4 +1,4 @@
-package soma.haeya.edupi_gateway;
+package soma.edupi.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class EdupiGatewayApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(EdupiGatewayApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(EdupiGatewayApplication.class, args);
+    }
 
 }
