+++ src/main/java/soma/edupi/gateway/config/RoutingConfig.java
@@ -1,11 +1,11 @@
-package soma.haeya.edupi_gateway.config;
+package soma.edupi.gateway.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.gateway.route.RouteLocator;
 import org.springframework.cloud.gateway.route.builder.RouteLocatorBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import soma.haeya.edupi_gateway.filter.AuthenticationFilter;
+import soma.edupi.gateway.filter.AuthenticationFilter;
 
 @Configuration
 public class RoutingConfig {
