+++ src/test/java/soma/edupi/gateway/EdupiGatewayApplicationTests.java
@@ -1,4 +1,4 @@
-package soma.haeya.edupi_gateway;
+package soma.edupi.gateway;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class EdupiGatewayApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
